requirements.txt
pyodbc>=4.0.35
confluent-kafka[avro]>=2.3.0
pyyaml>=6.0
pydantic>=2.0
schedule>=1.1.0
tenacity>=8.2.0
click>=8.1.0
structlog>=23.1.0
prometheus-client>=0.16.0